Better Repo Organization
Issue #3: In order to better enable contributions to this, I
extracted the contents of the original archive file and moved it
to the dist/ directory, updating the readme for the updated
location.

I also made some additional formatting changes to the readme and
renamed the tarball to indicate the first release numbering.
The installation instructions also has the SHA512sum of the
tarball echoed into sha512sum --check for download verification
purposes.

Hopefully the liberties I took aren't too great. Feel free to
reject this pull request if I'm overstepping here.